backend/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1
django-allauth==0.42.0
django-cors-headers==3.4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.10
mysqlclient==2.0.1
oauthlib==3.1.0
Pillow==7.2.0
psycopg2-binary==2.8.3
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
